backend/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.22
pydantic>=2.0
pydantic-settings>=2.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9
itsdangerous>=2.2.0
alembic>=1.11
python-dotenv>=1.0
redis>=5.0
celery>=5.3
cryptography>=41.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
openai>=1.0
httpx>=0.24
loguru>=0.7
textstat>=0.7.4
pyotp>=2.9.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
email-validator>=2.1.0